MR: remove Hive dependencies on Iceberg de/serialization utility functions
