+++ mr/src/main/java/org/apache/iceberg/mr/hive/HiveIcebergStorageHandler.java
@@ -20,7 +20,6 @@
 package org.apache.iceberg.mr.hive;
 
 import java.io.Serializable;
-import java.util.Collection;
 import java.util.Map;
 import java.util.Properties;
 import org.apache.hadoop.conf.Configuration;
@@ -36,19 +35,16 @@ import org.apache.hadoop.hive.serde2.Deserializer;
 import org.apache.hadoop.mapred.InputFormat;
 import org.apache.hadoop.mapred.JobConf;
 import org.apache.hadoop.mapred.OutputFormat;
-import org.apache.iceberg.Schema;
 import org.apache.iceberg.SchemaParser;
 import org.apache.iceberg.Table;
 import org.apache.iceberg.mr.Catalogs;
 import org.apache.iceberg.mr.InputFormatConfig;
 import org.apache.iceberg.relocated.com.google.common.annotations.VisibleForTesting;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
-import org.apache.iceberg.relocated.com.google.common.base.Splitter;
 import org.apache.iceberg.relocated.com.google.common.collect.Maps;
 import org.apache.iceberg.util.SerializationUtil;
 
 public class HiveIcebergStorageHandler implements HiveStoragePredicateHandler, HiveStorageHandler {
-  private static final Splitter TABLE_NAME_SPLITTER = Splitter.on("..");
   private static final String TABLE_NAME_SEPARATOR = "..";
 
   static final String WRITE_KEY = "HiveIcebergStorageHandler_write";
@@ -155,44 +151,6 @@ public class HiveIcebergStorageHandler implements HiveStoragePredicateHandler, H
     predicate.residualPredicate = (ExprNodeGenericFuncDesc) exprNodeDesc;
     predicate.pushedPredicate = (ExprNodeGenericFuncDesc) exprNodeDesc;
     return predicate;
-  }
-
-  /**
-   * Returns the Table serialized to the configuration based on the table name.
-   * @param config The configuration used to get the data from
-   * @param name The name of the table we need as returned by TableDesc.getTableName()
-   * @return The Table
-   */
-  public static Table table(Configuration config, String name) {
-    return SerializationUtil.deserializeFromBase64(config.get(InputFormatConfig.SERIALIZED_TABLE_PREFIX + name));
-  }
-
-  /**
-   * Returns the names of the output tables stored in the configuration.
-   * @param config The configuration used to get the data from
-   * @return The collection of the table names as returned by TableDesc.getTableName()
-   */
-  public static Collection<String> outputTables(Configuration config) {
-    return TABLE_NAME_SPLITTER.splitToList(config.get(InputFormatConfig.OUTPUT_TABLES));
-  }
-
-  /**
-   * Returns the catalog name serialized to the configuration.
-   * @param config The configuration used to get the data from
-   * @param name The name of the table we neeed as returned by TableDesc.getTableName()
-   * @return catalog name
-   */
-  public static String catalogName(Configuration config, String name) {
-    return config.get(InputFormatConfig.TABLE_CATALOG_PREFIX + name);
-  }
-
-  /**
-   * Returns the Table Schema serialized to the configuration.
-   * @param config The configuration used to get the data from
-   * @return The Table Schema object
-   */
-  public static Schema schema(Configuration config) {
-    return SchemaParser.fromJson(config.get(InputFormatConfig.TABLE_SCHEMA));
   }
 
   /**
